fix: :bug: assignment of multiple expressions now works
Before, the assigment of each value in an assigment statement was
processed separately. That meant that a simple swap operation in
assignment would not work, as the value would be overwriten. Now the
temporaries are kept until the end of the statement, ensuring that the
values live for the expected ammount of time.